requirements.txt
asgiref==3.2.7
astroid==2.4.1
boto3==1.14.19
botocore==1.17.19
certifi==2020.11.8
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
Django==2.1.1
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.10
isort==4.3.21
jmespath==0.10.0
mccabe==0.6.1
npm==0.1.1
optional-django==0.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
